+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import Board from "./Board";
+import { Board } from "./components";
 
 const App: React.FC = () => {
   return (
