+++ src/components/index.js
@@ -1,4 +1,5 @@
+import Board from "./Board";
 import Tile from "./Tile";
 import ColorSelect from "./ColorSelect";
 
-export { Tile, ColorSelect };
+export { Board, Tile, ColorSelect };
